Refactors DPC parameters into a parameters folder
