+++ .integration/tests/dpc_testnet1.rs
@@ -20,7 +20,7 @@ use snarkvm_algorithms::{
     SRS,
 };
 use snarkvm_curves::bls12_377::{Fq, Fr};
-use snarkvm_dpc::{prelude::*, testnet1::parameters::*};
+use snarkvm_dpc::{prelude::*, testnet1::*};
 use snarkvm_integration::{testnet1::*, *};
 use snarkvm_ledger::prelude::*;
 use snarkvm_r1cs::{ConstraintSystem, TestConstraintSystem};
