+++ parameters/examples/noop_program_snark.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_dpc::{testnet1::parameters::Testnet1Parameters, DPCError, NoopProgram, Parameters, ProgramScheme};
+use snarkvm_dpc::{testnet1::Testnet1Parameters, DPCError, NoopProgram, Parameters, ProgramScheme};
 use snarkvm_utilities::ToBytes;
 
 use rand::thread_rng;
