+++ parameters/examples/testnet2_universal_srs.rs
@@ -15,7 +15,7 @@
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
 use snarkvm_algorithms::{crh::sha256::sha256, SNARK};
-use snarkvm_dpc::{errors::DPCError, testnet2::parameters::Testnet2Parameters, Parameters};
+use snarkvm_dpc::{errors::DPCError, testnet2::Testnet2Parameters, Parameters};
 use snarkvm_utilities::ToBytes;
 
 use rand::thread_rng;
